+++ mindsdb/__main__.py
@@ -22,6 +22,7 @@ from typing import Callable, Optional, Tuple, List
 from sqlalchemy.orm.attributes import flag_modified
 
 from mindsdb.__about__ import __version__ as mindsdb_version
+from mindsdb.utilities.config import Config
 from mindsdb.api.http.start import start as start_http
 from mindsdb.api.mysql.start import start as start_mysql
 from mindsdb.api.mongo.start import start as start_mongo
@@ -29,7 +30,6 @@ from mindsdb.api.postgres.start import start as start_postgres
 from mindsdb.interfaces.tasks.task_monitor import start as start_tasks
 from mindsdb.utilities.ml_task_queue.consumer import start as start_ml_task_queue
 from mindsdb.interfaces.jobs.scheduler import start as start_scheduler
-from mindsdb.utilities.config import Config
 from mindsdb.utilities.ps import is_pid_listen_port, get_child_pids
 from mindsdb.utilities.functions import args_parse, get_versions_where_predictors_become_obsolete
 from mindsdb.interfaces.database.integrations import integration_controller
